Moved the discipline constants to utils/lib
